Fix bug when badge is changed back to nothing, remove unnecessary timer
Prior to this, if the badge icon was showing some data and then set to off/nothing, it would continue showing the old data.

Additionally, the background alarm would be set even if the badge was set to off/nothing. It now only configures the alarm when the badge needs to be updated.

Fixes https://github.com/myfreeweb/transmitter/issues/31